+++ bee-common-derive/tests/packable/fail/dup_id_enum.rs
@@ -1,10 +1,18 @@
 // Copyright 2021 IOTA Stiftung
 // SPDX-License-Identifier: Apache-2.0
 
+#![allow(unused_imports)]
+
 use bee_common::packable::Packable;
 use bee_common_derive::Packable;
 
 #[derive(Packable)]
-pub struct Num(u32);
+#[packable(ty = u8)]
+pub enum OptI32 {
+    #[packable(id = 0)]
+    None,
+    #[packable(id = 0)]
+    Some(i32)
+}
 
 fn main() {}
