+++ bee-common-derive/tests/packable/fail/invalid_ty_enum.rs
@@ -2,16 +2,17 @@
 // SPDX-License-Identifier: Apache-2.0
 
 #![allow(unused_imports)]
+
 use bee_common::packable::Packable;
 use bee_common_derive::Packable;
 
 #[derive(Packable)]
-#[packable(ty = u8)]
-pub enum OptPoint {
+#[packable(ty = [u8; 32])]
+pub enum OptI32 {
     #[packable(id = 0)]
     None,
     #[packable(id = 1)]
-    Some{ x: i32, y: i32 }
+    Some(i32)
 }
 
 fn main() {}
