+++ bee-common-derive/tests/packable/fail/no_id_enum.rs
@@ -2,10 +2,15 @@
 // SPDX-License-Identifier: Apache-2.0
 
 #![allow(unused_imports)]
+
 use bee_common::packable::Packable;
 use bee_common_derive::Packable;
 
 #[derive(Packable)]
-pub struct Wrap<T>(T);
+#[packable(ty = i32)]
+pub enum OptI32 {
+    None,
+    Some(i32)
+}
 
 fn main() {}
