+++ bee-common-derive/tests/packable/fail/no_id_enum.stderr
@@ -1,5 +1,5 @@
 error: All variants of an enum that derives `Packable` require a `#[packable(id = ...)]` attribute.
-  --> $DIR/no_id_enum.rs:11:5
+  --> $DIR/no_id_enum.rs:12:5
    |
-11 |     None,
+12 |     None,
    |     ^^^^
