+++ bee-common-derive/tests/packable/fail/union.rs
@@ -2,13 +2,14 @@
 // SPDX-License-Identifier: Apache-2.0
 
 #![allow(unused_imports)]
+
 use bee_common::packable::Packable;
 use bee_common_derive::Packable;
 
 #[derive(Packable)]
-pub struct Point {
-    x: i32,
-    y: i32,
+pub union MaybeInt {
+    just: u32,
+    nothing: (),
 }
 
 fn main() {}
