+++ bee-common-derive/tests/packable/pass/generic_newtype.rs
@@ -2,10 +2,11 @@
 // SPDX-License-Identifier: Apache-2.0
 
 #![allow(unused_imports)]
+
 use bee_common::packable::Packable;
 use bee_common_derive::Packable;
 
 #[derive(Packable)]
-pub struct Point(i32, i32);
+pub struct Wrap<T>(T);
 
 fn main() {}
