+++ bee-common-derive/tests/packable/pass/mixed_enum.rs
@@ -2,16 +2,17 @@
 // SPDX-License-Identifier: Apache-2.0
 
 #![allow(unused_imports)]
+
 use bee_common::packable::Packable;
 use bee_common_derive::Packable;
 
 #[derive(Packable)]
 #[packable(ty = u8)]
-pub enum OptI32 {
+pub enum Foo {
     #[packable(id = 0)]
-    None,
-    #[packable(id = 0)]
-    Some(i32)
+    Bar(u32),
+    #[packable(id = 1)]
+    Baz{ x: i32, y: i32 }
 }
 
 fn main() {}
