+++ bee-common-derive/tests/packable/pass/unnamed_struct.rs
@@ -2,14 +2,11 @@
 // SPDX-License-Identifier: Apache-2.0
 
 #![allow(unused_imports)]
+
 use bee_common::packable::Packable;
 use bee_common_derive::Packable;
 
 #[derive(Packable)]
-#[packable(ty = i32)]
-pub enum OptI32 {
-    None,
-    Some(i32)
-}
+pub struct Point(i32, i32);
 
 fn main() {}
