--- bee-common-derive/tests/packable/union.stderr
@@ -1,5 +1,0 @@
-error: Unions cannot derive `Packable`
- --> $DIR/union.rs:9:11
-  |
-9 | pub union MaybeInt {
-  |           ^^^^^^^^
